+++ contracts/interfaces/IGoosebumpsRouterPair.sol
@@ -5,5 +5,5 @@ pragma solidity 0.8.7;
 import "./IGoosebumpsPair.sol";
 
 interface IGoosebumpsRouterPair is IGoosebumpsPair {
-    function swap(uint amount0Out, uint amount1Out, address to) external;
+    function swap(uint256 amount0Out, uint256 amount1Out, address to) external;
 }
